SANY: Start to sort out syntax parser error reporting Emit PlusCal divergence warnings through Errors class
Currently, error reporting in SANY's first stage (the syntax parser) is
a bit of a mess. The current state of things:
 1. The TLAplusParser class throws ParseException instances, which are
    caught and turned into AbortException instances, which are caught
    and turned back into ParseException instances. The error message is
    output to the log but loses its attachment to the exceptions along
    the way.
 2. PlusCal translation divergence warnings are output directly to the
    log instead of tagged with a standardized error code and stored in
    the Errors class like other anomalous occurrences.
 3. Warning and error messages stored in the Errors class are both
    output at the Errors log level, instead of separated appropriately.

These changes currently do not really touch the first issue, but handle
the second and third. Another error message is also tagged with a
best-effort location.

The second and third stages (semantic & level-checking) are in better
shape but I modified them to re-use the error log presentation logic
from the first phase.

[SANY][Refactor]

Signed-off-by: Andrew Helwer <ahelwer@pm.me>